+++ R/beesTaxonomy.R
@@ -68,8 +68,9 @@
 #'  **notes** Additional notes about the name/taxon.
 #'  
 #'  **Previous taxonomies:**
+#'   - 2026-05-04 **current**: https://api.figshare.com/v2/file/download/64193731
 #'  
-#'   - 2026-01-12 **current**: https://open.flinders.edu.au/ndownloader/files/60945820
+#'   - 2026-01-12: https://open.flinders.edu.au/ndownloader/files/60945820
 #'  
 #'   - 2024-06-17: https://open.flinders.edu.au/ndownloader/files/47089969
 #'   
@@ -116,7 +117,7 @@
 #'}
 #' 
 #'
-beesTaxonomy <- function(URL = "https://api.figshare.com/v2/file/download/60945820",
+beesTaxonomy <- function(URL = "https://api.figshare.com/v2/file/download/64193731",
                          mode = NULL,
                          headers = NULL,
                          token = NULL,